requirements.txt
flask
flask-cors
tensorflow==2.12.0
pillow
numpy
gunicorn
